Moved components used by settings to the settings component folder
Moved components that the settings page uses to the settings component folder. It just makes sense and has less folder clutter in the component folder
